Updated unneeded tests

+++ iam/src/test/java/uk/ac/ox/oucs/iam/security/keys/CryptServicesTest.java
@@ -22,21 +22,22 @@ import uk.ac.ox.oucs.iam.security.utilities.GeneralUtils;
 import uk.ac.ox.oucs.iam.security.utilities.exceptions.KeyNotFoundException;
 import uk.ac.ox.oucs.iam.security.utilities.exceptions.NoEncodingException;
 
-
-
 /**
+ * NOTE: Currently the class this tests is not needed, so tests disabled
+ * 
  * Inputs: cryptkey.source - symmetric key generated on a remote machine
- * encString0-5 - encrypted messages from the remote machine that we need to decrypt
- * permanentSymKey - a file created during the first run of the tests and then used in subsequent runs
+ * encString0-5 - encrypted messages from the remote machine that we need to
+ * decrypt permanentSymKey - a file created during the first run of the tests
+ * and then used in subsequent runs
  * 
  * @author oucs0153
- *
+ * 
  */
 public class CryptServicesTest {
+	private static boolean disabled = true;
 	private static String[] msg;
 	private static int numberOfStrings = 0;
 	private static File permanentKey, remoteKey, localKey;
-	
 
 	@BeforeClass
 	public static void initialise() {
@@ -52,50 +53,29 @@ public class CryptServicesTest {
 		remoteKey = new File("testfiles" + File.separator + "cryptkey.source");
 		localKey = new File("localKey");
 	}
-	
-	
+
 	/**
-	 * Check we can decrypt remotely prepared encrypted strings
+	 * Check we cannot decrypt remotely prepared encrypted strings with a newly
+	 * created key
+	 * 
+	 * @throws NoEncodingException
+	 * @throws KeyNotFoundException
+	 * @throws IOException
+	 * @throws NoSuchPaddingException
+	 * @throws NoSuchAlgorithmException
+	 * @throws InvalidKeyException
+	 * @throws IllegalBlockSizeException
+	 * @throws BadPaddingException
 	 */
-//	@Test
-//	public void decryptRemoteStrings() {
-//		int counter = 0;
-//		try {
-//			CryptServices cs = new CryptServices(
-//					remoteKey.getAbsolutePath(), false);
-//			for (String s : msg) {
-//				if (s.length() == 0) {
-//					break;
-//				}
-//				String toTest = (String) GeneralUtils
-//						.readObjectFromFile("testfiles" + File.separator + "encString" + counter);
-//				assertTrue(cs.decrypt(toTest).equals(msg[counter]));
-//				counter++;
-//			}
-//		} catch (Exception e) {
-//			assertTrue(false);
-//		}
-//	}
-	
-	/**
-	 * Check we cannot decrypt remotely prepared encrypted strings
-	 * with a newly created key 
-	 * @throws NoEncodingException 
-	 * @throws KeyNotFoundException 
-	 * @throws IOException 
-	 * @throws NoSuchPaddingException 
-	 * @throws NoSuchAlgorithmException 
-	 * @throws InvalidKeyException 
-	 * @throws IllegalBlockSizeException 
-	 * @throws BadPaddingException 
-	 */
-	@Test (expected=BadPaddingException.class)
-	public void decryptRemoteStringsWithLocalKey() throws InvalidKeyException, NoSuchAlgorithmException, NoSuchPaddingException, IOException, KeyNotFoundException, NoEncodingException, IllegalBlockSizeException, BadPaddingException {
-		if (!GlobalTestVars.runTests) {
+	@Test(expected = BadPaddingException.class)
+	public void decryptRemoteStringsWithLocalKey() throws InvalidKeyException, NoSuchAlgorithmException,
+			NoSuchPaddingException, IOException, KeyNotFoundException, NoEncodingException, IllegalBlockSizeException,
+			BadPaddingException {
+		if (disabled) {
 			throw new BadPaddingException();
 		}
 		System.out.println("decryptRemoteStringsWithLocalKey");
-		
+
 		int counter = 0;
 		CryptServices cs = new CryptServices(localKey.getAbsolutePath(), true);
 		for (String s : msg) {
@@ -108,7 +88,7 @@ public class CryptServicesTest {
 			String toTest = (String) GeneralUtils.readObjectFromFile(fileName);
 			System.out.println("\tTest if <" + toTest + "> decrypts to <" + msg[counter] + "> (it should not)");
 			assertFalse(cs.decrypt(toTest).equals(msg[counter]));
-			
+
 			counter++;
 		}
 	}
@@ -119,46 +99,49 @@ public class CryptServicesTest {
 	 * The next time through, it will use that key file, read the strings and
 	 * use the permanent key file to decrypt the strings to make sure they can
 	 * be persisted.
-	 * @throws NoEncodingException 
-	 * @throws KeyNotFoundException 
-	 * @throws IOException 
-	 * @throws NoSuchPaddingException 
-	 * @throws NoSuchAlgorithmException 
-	 * @throws InvalidKeyException 
-	 * @throws BadPaddingException 
-	 * @throws IllegalBlockSizeException 
+	 * 
+	 * @throws NoEncodingException
+	 * @throws KeyNotFoundException
+	 * @throws IOException
+	 * @throws NoSuchPaddingException
+	 * @throws NoSuchAlgorithmException
+	 * @throws InvalidKeyException
+	 * @throws BadPaddingException
+	 * @throws IllegalBlockSizeException
 	 */
 	@Test
-	public void checkPersistance() throws InvalidKeyException, NoSuchAlgorithmException, NoSuchPaddingException, IOException, KeyNotFoundException, NoEncodingException, IllegalBlockSizeException, BadPaddingException {
-		if (!GlobalTestVars.runTests) {
+	public void checkPersistance() throws InvalidKeyException, NoSuchAlgorithmException, NoSuchPaddingException,
+			IOException, KeyNotFoundException, NoEncodingException, IllegalBlockSizeException, BadPaddingException {
+		if (disabled) {
 			return;
 		}
 		System.out.println("checkPersistance");
-		CryptServices cs = new CryptServices(
-				permanentKey.getAbsolutePath(), false);
+		CryptServices cs = new CryptServices(permanentKey.getAbsolutePath(), false);
 		int count = 0;
 		for (String s : msg) {
-			if ( (s == null) || (s.length() == 0) ) {
+			if ((s == null) || (s.length() == 0)) {
 				System.out.println("\tTests finished fine");
 				break;
 			}
 			String fileToTest = "./symEnc" + count;
 			System.out.println("\tTest file " + fileToTest);
 			String toTest = (String) GeneralUtils.readObjectFromFile(fileToTest);
-			count++;				
+			count++;
 			assertTrue(cs.decrypt(toTest).equals(s));
 			System.out.println("\ttested string");
 		}
 		System.out.println("\tsuccess");
 	}
-	
-	private void setupTestingFiles() throws InvalidKeyException, NoSuchAlgorithmException, NoSuchPaddingException, IOException, KeyNotFoundException, NoEncodingException, IllegalBlockSizeException, BadPaddingException {
+
+	private void setupTestingFiles() throws InvalidKeyException, NoSuchAlgorithmException, NoSuchPaddingException,
+			IOException, KeyNotFoundException, NoEncodingException, IllegalBlockSizeException, BadPaddingException {
+		if (disabled) {
+			return;
+		}
 		/*
-		 * This is the first time through - set up persistance strings
-		 * to file
+		 * This is the first time through - set up persistance strings to file
 		 */
-		CryptServices cs = new CryptServices(
-				permanentKey.getAbsolutePath(), true);
+		CryptServices cs = new CryptServices(permanentKey.getAbsolutePath(), true);
 		int count = 0;
 		for (String s : msg) {
 			if (s == null) {
@@ -167,26 +150,25 @@ public class CryptServicesTest {
 			String enc = cs.encrypt(s);
 			GeneralUtils.writeObject("./symEnc" + count++, enc);
 		}
-		System.out
-				.println("\tIt appears this is the first tme through this test - a permanent key has been\n"
-						+ "\tcreated. Please now re-run the test. If you see this message again then there is a problem!");
+		System.out.println("\tIt appears this is the first tme through this test - a permanent key has been\n"
+				+ "\tcreated. Please now re-run the test. If you see this message again then there is a problem!");
 	}
 
-	
 	/**
-	 * @throws BadPaddingException 
-	 * @throws IllegalBlockSizeException 
-	 * @throws NoEncodingException 
-	 * @throws KeyNotFoundException 
-	 * @throws IOException 
-	 * @throws NoSuchPaddingException 
-	 * @throws NoSuchAlgorithmException 
-	 * @throws InvalidKeyException 
+	 * @throws BadPaddingException
+	 * @throws IllegalBlockSizeException
+	 * @throws NoEncodingException
+	 * @throws KeyNotFoundException
+	 * @throws IOException
+	 * @throws NoSuchPaddingException
+	 * @throws NoSuchAlgorithmException
+	 * @throws InvalidKeyException
 	 * 
 	 */
 	@Test
-	public void basicEncryptionTests() throws IllegalBlockSizeException, BadPaddingException, InvalidKeyException, NoSuchAlgorithmException, NoSuchPaddingException, IOException, KeyNotFoundException, NoEncodingException {
-		if (!GlobalTestVars.runTests) {
+	public void basicEncryptionTests() throws IllegalBlockSizeException, BadPaddingException, InvalidKeyException,
+			NoSuchAlgorithmException, NoSuchPaddingException, IOException, KeyNotFoundException, NoEncodingException {
+		if (disabled) {
 			return;
 		}
 		System.out.println("basicEncryptionTests");
@@ -204,13 +186,15 @@ public class CryptServicesTest {
 			System.out.println("\ttested string");
 		}
 	}
-	
-	
+
 	/**
 	 * Remove files generated for the tests
 	 */
 	@AfterClass
 	public static void finish() {
+		if (disabled) {
+			return;
+		}
 		new File("localKey").getAbsoluteFile().delete();
 		new File("cryptKey").getAbsoluteFile().delete();
 	}
